requirements.txt
cryptography
customtkinter
pillow
bcrypt
python-dotenv